# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarSharing API",
        "description": "CarSharing description",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/{id}/car/{car_id}": {
            "delete": {
                "tags": [
                    "cars"
                ],
                "summary": "Удаление машины у пользователя",
                "operationId": "20a51f1547b95b49bdb672b9f80f2225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
...
}